@@ -1,8 +1,9 @@
 {
   "name": "@stefanprobst/rehype-shiki",
-  "version": "1.1.1",
+  "version": "2.0.0",
   "main": "src/index.js",
   "types": "src/index.d.ts",
+  "type": "module",
   "repository": "git@github.com:stefanprobst/rehype-shiki.git",
   "author": "Stefan Probst <stefanprobst@kolabnow.com>",
   "license": "MIT",
@@ -14,44 +15,43 @@
     "format:fix": "yarn format --write",
     "lint": "eslint . --ignore-path .gitignore",
     "lint:fix": "yarn lint --fix",
-    "test": "jest"
-  },
-  "dependencies": {
-    "hast-util-raw": "^6.1.0",
-    "hast-util-to-string": "^1.0.4",
-    "unified": "^9.2.1",
-    "unist-util-visit": "^2.0.3"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "^12.0.1",
-    "@commitlint/config-conventional": "^12.0.1",
-    "@stefanprobst/commitlint-config": "^1.0.1",
-    "@stefanprobst/eslint-config-node": "^1.0.7",
-    "@stefanprobst/prettier-config": "^1.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
-    "eslint": "^7.22.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-import-resolver-typescript": "^2.4.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.3.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "husky": "^4.0.0",
-    "jest": "^26.6.3",
-    "lint-staged": "^10.5.4",
-    "prettier": "^2.2.1",
-    "rehype-parse": "^7.0.1",
-    "rehype-stringify": "^8.0.0",
-    "shiki": "^0.9.5",
-    "typescript": "^4.2.3"
-  },
-  "peerDependencies": {
-    "shiki": "^0.9.5"
+    "test": "NODE_OPTIONS='--experimental-vm-modules' jest"
   },
   "engines": {
-    "node": ">=12.9",
-    "yarn": "^1"
+    "node": "14.x",
+    "yarn": "1.x"
+  },
+  "dependencies": {
+    "hast-util-raw": "^7.1.1",
+    "hast-util-to-string": "^2.0.0",
+    "unified": "^10.1.0",
+    "unist-util-visit": "^4.0.0"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "^13.1.0",
+    "@stefanprobst/commitlint-config": "^1.0.2",
+    "@stefanprobst/eslint-config-node": "^2.0.1",
+    "@stefanprobst/prettier-config": "^1.0.5",
+    "@typescript-eslint/eslint-plugin": "^4.29.1",
+    "@typescript-eslint/parser": "^4.29.1",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-import-resolver-typescript": "^2.4.0",
+    "eslint-plugin-import": "^2.24.0",
+    "eslint-plugin-jest": "^24.4.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "husky": "^4.0.0",
+    "jest": "^27.0.6",
+    "lint-staged": "^11.1.2",
+    "prettier": "^2.3.2",
+    "rehype-parse": "^8.0.2",
+    "rehype-stringify": "^9.0.1",
+    "shiki": "^0.9.6",
+    "typescript": "^4.3.5"
+  },
+  "peerDependencies": {
+    "shiki": "^0.9.6"
   },
   "commitlint": {
     "extends": [
